feat: refactor PostalMessage + add @lombok
