+++ src/main/java/ca/mgamble/postal/api/PostalService.java
@@ -29,14 +29,16 @@ THE SOFTWARE.
 package ca.mgamble.postal.api;
 
 
-import ca.mgamble.postal.classes.APIResponse;
-import ca.mgamble.postal.classes.SendMessage;
-import ca.mgamble.postal.classes.SendRawMessage;
+import ca.mgamble.postal.api.message.PostalMessage;
+import ca.mgamble.postal.api.response.PostalApiResponse;
+import ca.mgamble.postal.api.response.SendRawMessage;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 
 import java.io.Closeable;
 import java.io.IOException;
+
+import lombok.Setter;
 import org.apache.log4j.Logger;
 import org.asynchttpclient.*;
 import java.util.concurrent.Future;
@@ -46,103 +48,97 @@ import java.util.logging.Level;
  *
  * @author mgamble
  */
-public class Postal implements Closeable {
-    
+public class PostalService implements Closeable {
+
+    private static final Logger logger = Logger.getLogger(PostalService.class);
+
     private static final String JSON = "application/json";
-    //   private final boolean closeClient;
     private final AsyncHttpClient client;
-    private final String url;
-    private final String apiKey;
-    private final Version version = new Version();
-    //  private final ObjectMapper mapper;
-    private final Logger logger;
+
+    @Setter
+    private String url;
+
+    @Setter
+    private String apiKey;
+
     private boolean closed = false;
-    Gson gson = new GsonBuilder().serializeNulls().create();
-    
-    public Postal(String url, String apiKey) {
-        this.logger = Logger.getLogger(Postal.class);
-        this.url = url;
+    private Gson gson = new GsonBuilder().serializeNulls().create();
+
+    public PostalService() {
         this.client = new DefaultAsyncHttpClient();
-        this.apiKey = apiKey;
-        
     }
 
-    //////////////////////////////////////////////////////////////////////
-    // Closeable interface methods
-    //////////////////////////////////////////////////////////////////////
+    public PostalService(String url, String apiKey) {
+        this();
+        this.url = url;
+        this.apiKey = apiKey;
+    }
+
     public boolean isClosed() {
         return closed || client.isClosed();
     }
-    
+
     public void close() {
         if (!client.isClosed()) {
             try {
                 client.close();
             } catch (IOException ex) {
-                java.util.logging.Logger.getLogger(Postal.class.getName()).log(Level.SEVERE, null, ex);
+                java.util.logging.Logger.getLogger(PostalService.class.getName()).log(Level.SEVERE, null, ex);
             }
         }
         closed = true;
     }
-    
-    public String getClientVersion() {
-        return version.getBuildNumber();
-    }
-    
-    public APIResponse sendRawmessage(SendRawMessage message) throws Exception {
-         Future<Response> f = client.executeRequest(buildRequest("POST", "send/raw", gson.toJson(message)));
+
+    public PostalApiResponse sendRawmessage(SendRawMessage message) throws Exception {
+        Future<Response> f = client.executeRequest(buildRequest("POST", "send/raw", gson.toJson(message)));
         Response r = f.get();
         if (r.getStatusCode() != 200) {
-            
+
             throw new Exception("Could not send raw message");
         } else {
-            return gson.fromJson(r.getResponseBody(), APIResponse.class);
-            
+            return gson.fromJson(r.getResponseBody(), PostalApiResponse.class);
+
         }
     }
-    public APIResponse sendMessage(SendMessage message) throws Exception {
-        
-        // Simple pre-flight checks 
-        if (message.getTo().size() > 50) { 
+    public PostalApiResponse sendMessage(PostalMessage message) throws Exception {
+
+        // Simple pre-flight checks
+        if (message.getTo().size() > 50) {
             throw new Exception("Too many recipients");
         }
-        if (message.getCc().size() > 50) { 
+        if (message.getCc().size() > 50) {
             throw new Exception("Too many CC contacts");
         }
-        if (message.getBcc().size() > 50) { 
+        if (message.getBcc().size() > 50) {
             throw new Exception("Too many BCC contacts");
         }
-        
-        Future<Response> f = client.executeRequest(buildRequest("POST", "send/message", gson.toJson(message)));
-        Response r = f.get();
-        if (r.getStatusCode() != 200) {
-            
+
+        Future<Response> future = client.executeRequest(buildRequest("POST", "send/message", gson.toJson(message)));
+        Response response = future.get();
+        if (response.getStatusCode() != 200) {
             throw new Exception("Could not send message");
         } else {
-            return gson.fromJson(r.getResponseBody(), APIResponse.class);
-            
+            return gson.fromJson(response.getResponseBody(), PostalApiResponse.class);
         }
     }
-    
+
     private Request buildRequest(String type, String subUrl) {
         RequestBuilder builder = new RequestBuilder(type);
-        Request request = builder.setUrl(this.url + "/api/v1/" + subUrl)
+        return builder.setUrl(this.url + "/api/v1/" + subUrl)
                 .addHeader("Accept", JSON)
                 .addHeader("Content-Type", JSON)
                 .addHeader("x-server-api-key", this.apiKey)
                 .build();
-        return request;
     }
-    
+
     private Request buildRequest(String type, String subUrl, String requestBody) {
         RequestBuilder builder = new RequestBuilder(type);
-        Request request = builder.setUrl(this.url + "/api/v1/" + subUrl)
+        return builder.setUrl(this.url + "/api/v1/" + subUrl)
                 .addHeader("Accept", JSON)
                 .addHeader("Content-Type", JSON)
                 .addHeader("x-server-api-key", this.apiKey)
                 .setBody(requestBody)
                 .build();
-        return request;
     }
-    
+
 }
