+++ src/main/java/ca/mgamble/postal/api/message/Header.java
@@ -1,5 +1,7 @@
 
-package ca.mgamble.postal.classes;
+package ca.mgamble.postal.api.message;
+
+import lombok.Data;
 
 /**
  *
@@ -29,56 +31,13 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
  */
 
-public class Attachment {
-    private String name;
-    private String content_type;
-    private String data;
+@Data
+public class Header {
+    private String key;
+    private String value;
 
-    
-    public Attachment(String name, String content_type, String data) {
-        this.name = name;
-        this.content_type = content_type;
-        this.data = data;
-    }
-    /**
-     * @return the name
-     */
-    public String getName() {
-        return name;
-    }
-
-    /**
-     * @param name the name to set
-     */
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    /**
-     * @return the content_type
-     */
-    public String getContent_type() {
-        return content_type;
-    }
-
-    /**
-     * @param content_type the content_type to set
-     */
-    public void setContent_type(String content_type) {
-        this.content_type = content_type;
-    }
-
-    /**
-     * @return the data
-     */
-    public String getData() {
-        return data;
-    }
-
-    /**
-     * @param data the data to set
-     */
-    public void setData(String data) {
-        this.data = data;
+    public Header(String key, String value) {
+        this.key = key;
+        this.value = value;
     }
 }
